simplify asset download component
