+++ src/pages/assets.tsx
@@ -13,32 +13,6 @@ import {
   useTheme,
 } from "@chakra-ui/react"
 
-// Assets
-import MergePandaSvg from "../assets/home/merge-panda.svg"
-import EthDiamondGlyph from "../assets/assets/eth-diamond-glyph.svg"
-import EthDiamondBlack from "../assets/assets/eth-diamond-black.svg"
-import EthDiamondRainbow from "../assets/assets/eth-diamond-rainbow.svg"
-import EthDiamondPurple from "../assets/assets/eth-diamond-purple.svg"
-import EthGlyphColored from "../assets/assets/eth-glyph-colored.svg"
-import EthPortraitBlack from "../assets/assets/ethereum-logo-portrait-black.svg"
-import EthLandscapeBlack from "../assets/assets/ethereum-logo-landscape-black.svg"
-import EthWordmarkBlack from "../assets/assets/ethereum-wordmark-black.svg"
-import EthPortraitPurple from "../assets/assets/ethereum-logo-portrait-purple-purple.svg"
-import EthLandscapePurple from "../assets/assets/ethereum-logo-landscape-purple.svg"
-import EthWordmarkPurple from "../assets/assets/ethereum-wordmark-purple-purple.svg"
-import EthDiamondBlackWhite from "../assets/assets/eth-diamond-black-white.svg"
-import EthDiamondBlackGrey from "../assets/assets/eth-diamond-black-gray.svg"
-import EthDiamondPurplePurple from "../assets/assets/eth-diamond-purple-purple.svg"
-import EthDiamondPurpleWhite from "../assets/assets/eth-diamond-purple-white.svg"
-import EthPortraitBlackGray from "../assets/assets/ethereum-logo-portrait-black-gray.svg"
-import EthLandscapeBlackGray from "../assets/assets/ethereum-logo-landscape-black-gray.svg"
-import EthWordmarkBlackGray from "../assets/assets/ethereum-wordmark-black-gray.svg"
-import EthPortraitPurplePurple from "../assets/assets/ethereum-logo-portrait-purple-purple.svg"
-import EthLandscapePurplePurple from "../assets/assets/ethereum-logo-landscape-purple-purple.svg"
-import EthWordmarkPurplePurple from "../assets/assets/ethereum-wordmark-purple-purple.svg"
-import EthLandscapePurpleWhite from "../assets/assets/ethereum-logo-landscape-purple-white.svg"
-import EthWordmarkPurpleWhite from "../assets/assets/ethereum-wordmark-purple-white.svg"
-
 // Components
 import AssetDownload from "../components/AssetDownload"
 import Link from "../components/Link"
@@ -149,11 +123,7 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             title={t("page-assets-merge-panda")}
             alt={t("page-assets-merge-panda")}
             image={data.mergePanda}
-          />
-          <AssetDownload
-            title={t("page-assets-merge-panda-svg")}
-            alt={t("page-assets-merge-panda-svg")}
-            svg={MergePandaSvg}
+            svgUrl="/static/merge-panda.svg"
           />
         </Row>
         <Row>
@@ -226,14 +196,6 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             artistName="William Tempest"
             artistUrl="https://cargocollective.com/willtempest"
           />
-          <AssetDownload
-            title={t("page-assets-robot")}
-            alt={t("page-assets-robot")}
-            image={data.wallet}
-            artistName="William Tempest"
-            artistUrl="https://cargocollective.com/willtempest"
-            hideBelow="lg"
-          />
         </Row>
         <Row>
           <AssetDownload
@@ -299,23 +261,6 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             artistUrl="https://www.patrickatkins.co.uk/"
           />
         </Row>
-        <Row>
-          <AssetDownload
-            title={t("page-assets-leslie-the-rhino")}
-            alt={t("page-assets-leslie-the-rhino")}
-            artistName="Tomo Saito"
-            artistUrl="https://tomosaito.com/"
-            image={data.leslieTheRhino}
-          />
-          <AssetDownload
-            title={t("page-assets-leslie-the-rhino")}
-            alt={t("page-assets-leslie-the-rhino")}
-            artistName="Tomo Saito"
-            artistUrl="https://tomosaito.com/"
-            image={data.leslieTheRhino}
-            hideBelow="lg"
-          />
-        </Row>
         <H2 id="historical">
           <Translation id="page-assets-historical-artwork" />
         </H2>
@@ -330,19 +275,19 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             title={t("page-assets-eth-diamond-glyph")}
             alt={t("page-assets-eth-diamond-glyph")}
             image={data.ethDiamondGlyph}
-            svg={EthDiamondGlyph}
+            svgUrl="/static/eth-diamond-glyph.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-diamond-gray")}
             alt={t("page-assets-eth-diamond-gray")}
             image={data.ethDiamondBlack}
-            svg={EthDiamondBlack}
+            svgUrl="/static/eth-diamond-black.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-diamond-color")}
             alt={t("page-assets-eth-diamond-color")}
             image={data.ethDiamondColor}
-            svg={EthDiamondRainbow}
+            svgUrl="/static/eth-diamond-rainbow.svg"
           />
         </Row>
         <Row>
@@ -350,13 +295,13 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             title={t("page-assets-eth-diamond-purple")}
             alt={t("page-assets-eth-diamond-purple")}
             image={data.ethDiamondPurple}
-            svg={EthDiamondPurple}
+            svgUrl="/static/eth-diamond-purple.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-diamond-colored")}
             alt={t("page-assets-eth-diamond-colored")}
             image={data.ethGlyphColored}
-            svg={EthGlyphColored}
+            svgUrl="/static/eth-glyph-colored.svg"
           />
         </Row>
         <Row>
@@ -364,19 +309,19 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             title={t("page-assets-eth-logo-portrait-gray")}
             alt={t("page-assets-eth-logo-portrait-gray")}
             image={data.ethPortraitBlack}
-            svg={EthPortraitBlack}
+            svgUrl="/static/ethereum-logo-portrait-black.svg  "
           />
           <AssetDownload
             title={t("page-assets-eth-logo-landscape-gray")}
             alt={t("page-assets-eth-logo-landscape-gray")}
             image={data.ethLandscapeBlack}
-            svg={EthLandscapeBlack}
+            svgUrl="/static/ethereum-logo-landscape-black.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-wordmark-gray")}
             alt={t("page-assets-eth-wordmark-gray")}
             image={data.ethWordmarkBlack}
-            svg={EthWordmarkBlack}
+            svgUrl="/static/ethereum-wordmark-black.svg"
           />
         </Row>
         <Row>
@@ -384,19 +329,19 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             title={t("page-assets-eth-logo-portrait-purple")}
             alt={t("page-assets-eth-logo-portrait-purple")}
             image={data.ethPortraitPurple}
-            svg={EthPortraitPurple}
+            svgUrl="/static/ethereum-logo-portrait-purple.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-logo-landscape-purple")}
             alt={t("page-assets-eth-logo-landscape-purple")}
             image={data.ethLandscapePurple}
-            svg={EthLandscapePurple}
+            svgUrl="/static/ethereum-logo-landscape-purple.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-wordmark-purple")}
             alt={t("page-assets-eth-wordmark-purple")}
             image={data.ethWordmarkPurple}
-            svg={EthWordmarkPurple}
+            svgUrl="/static/ethereum-wordmark-purple-purple.svg"
           />
         </Row>
         <H3>
@@ -407,19 +352,19 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             title={t("page-assets-eth-diamond-white")}
             alt={t("page-assets-eth-diamond-white")}
             image={data.ethDiamondBlackWhite}
-            svg={EthDiamondBlackWhite}
+            svgUrl="/static/eth-diamond-black-white.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-diamond-gray")}
             alt={t("page-assets-eth-diamond-gray")}
             image={data.ethDiamondBlackGray}
-            svg={EthDiamondBlackGrey}
+            svgUrl="/static/eth-diamond-black-gray.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-diamond-purple")}
             alt={t("page-assets-eth-diamond-purple")}
             image={data.ethDiamondPurplePurple}
-            svg={EthDiamondPurplePurple}
+            svgUrl="/static/eth-diamond-purple-purple.svg"
           />
         </Row>
 
@@ -428,27 +373,13 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             title={t("page-assets-eth-diamond-white")}
             alt={t("page-assets-eth-diamond-white")}
             image={data.ethDiamondPurpleWhite}
-            svg={EthDiamondPurpleWhite}
+            svgUrl="/static/eth-diamond-purple-white.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-diamond-white")}
             alt={t("page-assets-eth-diamond-white")}
             image={data.ethDiamondPurpleWhite}
-            svg={EthDiamondPurpleWhite}
-          />
-          <AssetDownload
-            title={t("page-assets-eth-diamond-white")}
-            alt={t("page-assets-eth-diamond-white")}
-            image={data.ethDiamondPurpleWhite}
-            svg={EthDiamondPurpleWhite}
-            hideBelow="lg"
-          />
-          <AssetDownload
-            title={t("page-assets-eth-diamond-white")}
-            alt={t("page-assets-eth-diamond-white")}
-            image={data.ethDiamondPurpleWhite}
-            svg={EthDiamondPurpleWhite}
-            hideBelow="lg"
+            svgUrl="/static/eth-diamond-purple-white.svg"
           />
         </Row>
         <Row>
@@ -456,19 +387,19 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             title={t("page-assets-eth-logo-portrait-gray")}
             alt={t("page-assets-eth-logo-portrait-gray")}
             image={data.ethPortraitBlackGray}
-            svg={EthPortraitBlackGray}
+            svgUrl="/static/ethereum-logo-portrait-black-gray.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-logo-landscape-gray")}
             alt={t("page-assets-eth-logo-landscape-gray")}
             image={data.ethLandscapeBlackGray}
-            svg={EthLandscapeBlackGray}
+            svgUrl="/static/ethereum-logo-landscape-black-gray.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-wordmark-gray")}
             alt={t("page-assets-eth-wordmark-gray")}
             image={data.ethWordmarkBlackGray}
-            svg={EthWordmarkBlackGray}
+            svgUrl="/static/ethereum-wordmark-black-gray.svg"
           />
         </Row>
         <Row>
@@ -476,39 +407,33 @@ const AssetsPage = ({ data }: PageProps<Queries.AssetsPageQuery, Context>) => {
             title={t("page-assets-eth-logo-portrait-purple")}
             alt={t("page-assets-eth-logo-portrait-purple")}
             image={data.ethPortraitPurplePurple}
-            svg={EthPortraitPurplePurple}
+            svgUrl="/static/ethereum-logo-portrait-purple-purple.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-logo-landscape-purple")}
             alt={t("page-assets-eth-logo-landscape-purple")}
             image={data.ethLandscapePurplePurple}
-            svg={EthLandscapePurplePurple}
+            svgUrl="/static/ethereum-logo-landscape-purple-purple.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-wordmark-purple")}
             alt={t("page-assets-eth-wordmark-purple")}
             image={data.ethWordmarkPurplePurple}
-            svg={EthWordmarkPurplePurple}
+            svgUrl="/static/ethereum-wordmark-purple-purple.svg"
           />
         </Row>
         <Row>
           <AssetDownload
-            title={t("page-assets-eth-logo-portrait-white")}
-            alt={t("page-assets-eth-logo-portrait-white")}
-            image={data.ethPortraitPurpleWhite}
-            svg={EthWordmarkPurplePurple}
-          />
-          <AssetDownload
             title={t("page-assets-eth-logo-landscape-white")}
             alt={t("page-assets-eth-logo-landscape-white")}
             image={data.ethLandscapePurpleWhite}
-            svg={EthLandscapePurpleWhite}
+            svgUrl="/static/ethereum-logo-landscape-purple-white.svg"
           />
           <AssetDownload
             title={t("page-assets-eth-wordmark-white")}
             alt={t("page-assets-eth-wordmark-white")}
             image={data.ethWordmarkPurpleWhite}
-            svg={EthWordmarkPurpleWhite}
+            svgUrl="/static/ethereum-wordmark-purple-white.svg"
           />
         </Row>
       </Box>
